Move incremental clustering into its owning layer
Amp-Thread-ID: https://ampcode.com/threads/T-01a02998-d8d2-7039-b6de-8c3025395257
