+++ agents/scope_ids.py
@@ -1,1 +1,3 @@
-ROOT_SCOPE_ID = "root"
+from clustering_ids import ROOT_SCOPE_ID
+
+__all__ = ["ROOT_SCOPE_ID"]
