+++ diagram_analysis/exceptions.py
@@ -1,40 +1,5 @@
 """Exceptions raised by diagram_analysis pipelines."""
 
-from pathlib import Path
-
-from static_analyzer.analysis_cache import STATIC_ANALYSIS_PKL, STATIC_ANALYSIS_SHA
-
-
-class IncrementalCacheMissingError(RuntimeError):
-    """Raised when ``generate_analysis_incremental`` finds no usable warm cache.
-
-    Needs a populated ``ClusterCache`` on the cached ``LanguageResults``, from the
-    SHA-tagged ``static_analysis.pkl``. Why raise: falling back to a full analysis
-    discarded analysis.json's depth and component IDs. The message names which
-    piece is missing — pkl, sha, or cluster baseline.
-    """
-
-    def __init__(self, artifact_dir: Path, reason: str = ""):
-        pkl_path = artifact_dir / STATIC_ANALYSIS_PKL
-        sha_path = artifact_dir / STATIC_ANALYSIS_SHA
-        if not reason:
-            if not pkl_path.exists():
-                reason = f"no {STATIC_ANALYSIS_PKL} at {pkl_path}"
-            elif not sha_path.exists():
-                reason = (
-                    f"{STATIC_ANALYSIS_PKL} at {pkl_path} has no sibling "
-                    f"{STATIC_ANALYSIS_SHA}; the warm-start cannot SHA-gate"
-                )
-            else:
-                reason = (
-                    f"{STATIC_ANALYSIS_PKL} at {pkl_path} loaded but has no cluster baseline "
-                    "(legacy pkl or first-ever incremental run)"
-                )
-        super().__init__(
-            f"Incremental analysis cannot proceed: {reason}. " "Run a full analysis first to seed the cache."
-        )
-        self.artifact_dir = artifact_dir
-
 
 class IncrementalClusteringError(RuntimeError):
     """Raised when a scope's clustering comes back empty but it still owns live methods.
