+++ static_analyzer/clustering/delta.py
@@ -19,21 +19,12 @@ from pathlib import Path
 
 import networkx as nx
 
-from agents.content_hash import (
-    MethodSpan,
-    SourceCache,
-    hash_file_residual,
-    hash_method_body,
-    hash_whole_file,
-    read_source_lines,
-)
-from agents.file_index_models import FileEntry
-from diagram_analysis.cluster_snapshot import ClusterSnapshot, ClusterSnapshotEntry
-from repo_utils.path_utils import normalize_repo_path
 from repo_utils.change_detector import ChangeSet
+from repo_utils.path_utils import normalize_repo_path
 from static_analyzer.cfg import DEFAULT_REFERENCE_KINDS
 from static_analyzer.analysis_result import StaticAnalysisResults
-from static_analyzer.clustering import ClusterResult
+from static_analyzer.clustering.models import ClusterResult
+from static_analyzer.clustering.snapshot import ClusterSnapshot, ClusterSnapshotEntry
 from static_analyzer.leiden_utils import find_partition_seeded
 
 logger = logging.getLogger(__name__)
@@ -62,140 +53,6 @@ class ClusterDelta:
 
     def cluster_results(self) -> dict[str, ClusterResult]:
         return {lang: d.cluster_results for lang, d in self.by_language.items()}
-
-
-@dataclass
-class ChangedMembers:
-    """Member-granular content-change signal from per-method content hashes.
-
-    ``members`` are qualified names (the identity clusters store, so they join
-    directly against cluster members) whose method body content changed — or
-    that were added/removed — inside the change set's added/modified files.
-    ``unattributed_files`` are changed files whose edit no hashed member
-    represents (module-level statements, data/config files, or source with no
-    indexed methods); the narrow file-level fallback dirties clusters owning
-    them so such edits are never silently missed.
-    """
-
-    # Bare qnames, not (file, qname): downstream dirty/restore checks match on qname alone, which
-    # assumes a qname is unique per file (module-path-prefixed qnames guarantee this). A cross-file
-    # bare-qname collision would let one file's edit dirty the other's cluster — accepted as a
-    # redundant re-detail, never a missed change.
-    members: set[str] = field(default_factory=set)
-    unattributed_files: set[str] = field(default_factory=set)
-
-
-def compute_changed_members(
-    baseline_files: dict[str, FileEntry],
-    new_static: StaticAnalysisResults,
-    changes: ChangeSet,
-    repo_dir: Path,
-    source_cache: SourceCache | None = None,
-) -> ChangedMembers:
-    """Diff per-method content hashes to find genuinely-changed cluster members.
-
-    ``baseline_files`` is the prior ``analysis.json`` file index (its methods
-    carry the previously-persisted ``content_hash``); the new hashes are
-    recomputed from live source at the current CFG spans with the same helper,
-    so an unchanged method below an edit — whose line span shifts but whose body
-    text is identical — hashes equal and does not light up. Scoped to the change
-    set's added/modified files, so drift in untouched files cannot leak in.
-
-    A changed file that produces no changed member (and whose whole-file hash
-    still differs) is recorded in ``unattributed_files`` for the file-level
-    fallback. When the baseline predates content hashing (all hashes ``''``),
-    every method in a changed file differs and the signal degrades gracefully to
-    the old file-granular behavior rather than missing the change.
-    """
-    file_cache: SourceCache = source_cache if source_cache is not None else {}
-    changed_paths = {normalize_repo_path(fc.file_path, repo_dir) for fc in changes.files if fc.is_content_change()}
-    if not changed_paths:
-        return ChangedMembers()
-
-    new_member_hashes, new_member_spans = _live_member_hashes(new_static, repo_dir, changed_paths, file_cache)
-
-    result = ChangedMembers()
-    for path in changed_paths:
-        baseline_entry = baseline_files.get(path)
-        baseline_members = (
-            {method.qualified_name: method.content_hash for method in baseline_entry.methods}
-            if baseline_entry is not None
-            else {}
-        )
-        new_members = new_member_hashes.get(path, {})
-
-        file_changed: set[str] = set()
-        for qname in set(baseline_members) | set(new_members):
-            in_baseline = qname in baseline_members
-            in_new = qname in new_members
-            if in_baseline and in_new:
-                if baseline_members[qname] != new_members[qname]:
-                    file_changed.add(qname)
-            else:
-                # Present in exactly one index: a method was added or removed.
-                file_changed.add(qname)
-
-        result.members |= file_changed
-
-        # Module-level (non-method) content is not covered by any member hash, so a
-        # constant/import/decorator edit must be caught separately — even when a sibling
-        # method in the same file also changed (a "mixed" edit). Compare the residual of
-        # everything outside the live method spans against the baseline residual; only a
-        # real module-level difference dirties the file, so a pure member edit never does.
-        baseline_module_hash = baseline_entry.module_hash if baseline_entry is not None else ""
-        new_module_hash = hash_file_residual(
-            read_source_lines(repo_dir, path, file_cache), new_member_spans.get(path, [])
-        )
-        module_changed = new_module_hash != baseline_module_hash
-
-        if file_changed:
-            if module_changed and baseline_module_hash:
-                # Only trust the residual when the baseline actually carried one; a legacy
-                # baseline (no module_hash) would otherwise dirty the file on every mixed edit.
-                result.unattributed_files.add(path)
-            continue
-
-        # No hashed member represents this file's edit — fall back to file level,
-        # but only if the whole-file content actually differs (a fingerprint
-        # false positive or metadata-only change must not dirty the cluster).
-        baseline_file_hash = baseline_entry.content_hash if baseline_entry is not None else ""
-        new_file_hash = hash_whole_file(read_source_lines(repo_dir, path, file_cache))
-        if new_file_hash != baseline_file_hash:
-            result.unattributed_files.add(path)
-    return result
-
-
-def _live_member_hashes(
-    new_static: StaticAnalysisResults,
-    repo_dir: Path,
-    changed_paths: set[str],
-    file_cache: SourceCache,
-) -> tuple[dict[str, dict[str, str]], dict[str, list[MethodSpan]]]:
-    """``({file_path -> {qname -> body_hash}}, {file_path -> [method spans]})`` for changed files.
-
-    Data nodes are skipped so this population matches the persisted file index, which
-    ``build_files_index`` fills from component-owned methods only. A module-level variable
-    lives in the CFG but never in the index, so counting it here would report it as a
-    brand-new method on every run and — because its span would be excised from the
-    module-level residual that the index never excised — make ``module_hash`` differ from
-    the baseline for any file that has one, permanently. An edit to such a variable still
-    surfaces: its lines stay in the residual, so the module hash moves.
-    """
-    hashes: dict[str, dict[str, str]] = {}
-    spans: dict[str, list[MethodSpan]] = {}
-    for language in new_static.get_languages():
-        try:
-            cfg = new_static.get_cfg(language)
-        except (KeyError, ValueError):
-            continue
-        for qname, node in cfg.nodes.items():
-            path = normalize_repo_path(node.file_path, repo_dir)
-            if path not in changed_paths or node.is_data():
-                continue
-            source_lines = read_source_lines(repo_dir, path, file_cache)
-            hashes.setdefault(path, {})[qname] = hash_method_body(source_lines, node.line_start, node.line_end)
-            spans.setdefault(path, []).append(MethodSpan(node.line_start, node.line_end))
-    return hashes, spans
 
 
 def compute_cluster_delta(
@@ -229,7 +86,7 @@ def compute_cluster_delta(
         # re-cluster type-coupled methods differently and drift from what a full analysis produces.
         nx_graph = cfg.to_networkx(DEFAULT_REFERENCE_KINDS)
         old_clusters = old_snapshot.get_language(language)
-        language_delta = _delta_for_language(
+        language_delta = delta_for_language(
             language,
             nx_graph,
             old_clusters,
@@ -253,7 +110,7 @@ def _changeset_to_path_set(changes: ChangeSet) -> set[str]:
     return paths
 
 
-def _delta_for_language(
+def delta_for_language(
     language: str,
     nx_graph: nx.DiGraph,
     old_clusters: dict[int, ClusterSnapshotEntry],
