+++ tests/static_analyzer/test_call_graph.py
@@ -6,7 +6,8 @@ from static_analyzer.config import NodeType
 from static_analyzer.leiden_utils import find_partition
 from static_analyzer.node import Node
 from static_analyzer.cfg import CallGraph, DEFAULT_REFERENCE_KINDS, Edge
-from static_analyzer.clustering import ClusterResult, ClusteringService
+from static_analyzer.clustering import ClusterResult
+from static_analyzer.clustering.service import ClusteringService
 
 
 class TestNode(unittest.TestCase):
