+++ tests/static_analyzer/test_clustering.py
@@ -3,7 +3,8 @@ import unittest
 import networkx as nx
 
 from static_analyzer.cfg import CallGraph
-from static_analyzer.clustering import ClusterCache, ClusterResult, ClusteringService
+from static_analyzer.clustering import ClusterCache, ClusterResult
+from static_analyzer.clustering.service import ClusteringService
 from static_analyzer.config import ClusteringConfig
 from static_analyzer.leiden_utils import find_partition
 from static_analyzer.node import Node
